Split MPV Settings for all OS, added adpt.sharpen
Added Adaptive Sharpen, and removed CAS, as i think it works better than CAS most of the time.

The split of the different OS' Profiles was made due to errors reported in my GitHub Issues. I think these should be fixed, need to be tested as i havent had much time to do it.